requirements.txt
streamlit
pandas
json
google-generative-ai